Update logic that determines output path
Change logic to use sourceFileName from state.file.opts rather than filename. sourceFileName is relative to babel source root so it means that the use of process.cwd() is unnecessary. Since process.cwd() is not necessarily the same as source root it also makes the logic work in all cases - assuming that the babel config property sourceRoot is used.